pyproject.toml
[tool.poetry]
name = "electroservice"
version = "0.1.0"
description = "Test task for electroservice company"
authors = ["SergeiNaum <sergei.naumencko@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
flask-login = "^0.6.3"
flask-sqlalchemy = "^3.1.1"
wtforms = "^3.1.2"
flask-wtf = "^1.2.1"
pymysql = "^1.1.0"
flask-migrate = "^4.0.7"
flask-script = "^2.0.6"
email-validator = "^2.1.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
